Add allergen parsing support and integrate it into menu processing
- Introduced `AlergensParser` to process allergen bitmasks into individual indices.
- Updated `DailyMenu` and `Meal` models to store allergens as `Long` instead of `String/int`.
- Extended `DailyMenuMessageBuilder` to include parsed allergen data in menu outputs.
- Updated `JidloviceMenuClient` to correctly map allergen data from meals.
- Added unit tests to validate the correctness of allergen parsing logic.

+++ src/main/java/eu/kratochvil/jidlobot/client/dto/Meal.java
@@ -15,7 +15,7 @@ public class Meal {
     private String description_en;
     private double price;
     private Integer quantity;
-    private int allergens;
+    private long allergens;
     private boolean draft;
     private boolean permanent;
     private int id;
@@ -73,11 +73,11 @@ public class Meal {
         this.quantity = quantity;
     }
 
-    public int getAllergens() {
+    public long getAllergens() {
         return allergens;
     }
 
-    public void setAllergens(int allergens) {
+    public void setAllergens(long allergens) {
         this.allergens = allergens;
     }
